Always check the existence of auto tag for ready builds
The old version was workaround to allow rebuild the auto tagged commits.
However in git the translation of '.' and '@' in auto tag names is not
made. Together it could lead to wrong behaviour in the following
scenario:
- old style auto tag exists
- new style auto tag is requested for different commit that still
  represents the same version-release of the package

The better fix is to always check for existence of tag in ready builds,
and builder shouldn't fail when the tag is already present, but points
to the same commit as the one requested for build.